negative instance finding
This updates the current system to find negative instances for both training and evaluation and makes them the same. The code selects the top instances that have both never seen the token in their description and are the furthest away from the combined positive instances for that token. The proportion of instances to select is stored in the NEG_SAMPLE_PORTION variable.